fix: sentry 도메인 모듈로 이동(fix 필요)
