+++ domain/src/main/resources/db/migration/V202406100200__subscription_update_field.sql
