--- domain/src/main/resources/db/migration/V202406111654__payment_index_unique.sql
@@ -1,2 +1,0 @@
-ALTER TABLE `payment`
-    ADD UNIQUE INDEX `unique_payment_seq` (`subscription_id`,`seq`) USING BTREE;
