+++ domain/src/main/resources/sentry.properties
@@ -1,0 +1,4 @@
+#dsn=${SENTRY_DSN}
+#traces-sample-rate=1.0
+#environment=${ENVIRONMENT}
+#exception-resolver-order=-2147483647
